fix: use size_t instead of int to recieve file size to match the client
- Lord I am dumb as fuck
- No more bytes left on the socket that recv() is able to yoink away
- Still features SOME undefined behaviour I reckon but the meat of the issue is resolved
- Must also write the file in HOSTDIR rather than the PWD where the server is invoked

+++ src/server.c
@@ -200,7 +200,7 @@ int serv_wrap_upload(int cfd, char *buf) {
  * @param[buf] the buffer containing the request $UPLOAD$<filename>$
  */
 int serv_wrap_download(int cfd, char *buf) {
-	int fsize;
+	size_t fsize;
 	char *filename;
 
 	filename = buf + 8;
